Render tokens pages in main nav using standard markdown renderer
