+++ polaris.shopify.com/content/tokens/height.mdx
@@ -1,0 +1,19 @@
+---
+title: Height
+showTOC: false
+---
+
+# Tokens &rarr; {frontmatter.title}
+
+<TokensNav selected="height" />
+
+<TokenList>
+  {tokens.height
+    .map((token) => (
+      <TokenList.Item
+        key={token.name}
+        category="height"
+        token={token}
+      />
+    ))}
+</TokenList>
