+++ polaris.shopify.com/package.json
@@ -40,6 +40,7 @@
     "@juggle/resize-observer": "^3.4.0",
     "lodash.get": "^4.4.2",
     "lodash.throttle": "^4.1.1",
+    "lodash.mapvalues": "^4.6.0",
     "nanoid": "^4.0.2",
     "next": "^13.0.6",
     "next-mdx-remote": "^4.4.1",
@@ -67,6 +68,7 @@
     "@types/js-yaml": "^4.0.5",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.throttle": "^4.1.7",
+    "@types/lodash.mapvalues": "^4.6.9",
     "@types/marked": "^4.0.3",
     "@types/prismjs": "^1.26.0",
     "@types/react": "^18.2.0",
