--- polaris.shopify.com/pages/tokens/color.tsx
@@ -1,16 +1,0 @@
-import type {NextPage} from 'next';
-import React from 'react';
-import TokensPage from '../../src/components/TokensPage';
-import PageMeta from '../../src/components/PageMeta';
-
-const Components: NextPage = () => {
-  return (
-    <>
-      <PageMeta title="Color tokens" />
-
-      <TokensPage tokenGroup={'color'} />
-    </>
-  );
-};
-
-export default Components;
