--- polaris.shopify.com/pages/tokens/font.tsx
@@ -1,16 +1,0 @@
-import type {NextPage} from 'next';
-import React from 'react';
-import TokensPage from '../../src/components/TokensPage';
-import PageMeta from '../../src/components/PageMeta';
-
-const Components: NextPage = () => {
-  return (
-    <>
-      <PageMeta title="Font tokens" />
-
-      <TokensPage tokenGroup={'font'} />
-    </>
-  );
-};
-
-export default Components;
