--- polaris.shopify.com/pages/tokens/width.tsx
@@ -1,16 +1,0 @@
-import type {NextPage} from 'next';
-import React from 'react';
-import TokensPage from '../../src/components/TokensPage';
-import PageMeta from '../../src/components/PageMeta';
-
-const Components: NextPage = () => {
-  return (
-    <>
-      <PageMeta title="Width tokens" />
-
-      <TokensPage tokenGroup={'width'} />
-    </>
-  );
-};
-
-export default Components;
