+++ polaris.shopify.com/src/components/TokensNav/TokensNav.tsx
@@ -1,0 +1,95 @@
+import styles from './TokensNav.module.scss';
+import Link from 'next/link';
+import {slugify} from '../../utils/various';
+
+interface Props {
+  selected?:
+    | 'border'
+    | 'breakpoints'
+    | 'color'
+    | 'font'
+    | 'height'
+    | 'motion'
+    | 'shadow'
+    | 'space'
+    | 'text'
+    | 'width'
+    | 'zIndex';
+}
+
+export type NavItem = {
+  title: string;
+  url?: string;
+};
+
+const navItems: NavItem[] = [
+  {
+    title: 'Border',
+    url: `/tokens/border`,
+  },
+  {
+    title: 'Breakpoints',
+    url: `/tokens/breakpoints`,
+  },
+  {
+    title: 'Color',
+    url: `/tokens/color`,
+  },
+  {
+    title: 'Font',
+    url: `/tokens/font`,
+  },
+  {
+    title: 'Height',
+    url: `/tokens/height`,
+  },
+  {
+    title: 'Motion',
+    url: `/tokens/motion`,
+  },
+  {
+    title: 'Shadow',
+    url: `/tokens/shadow`,
+  },
+  {
+    title: 'Space',
+    url: `/tokens/space`,
+  },
+  {
+    title: 'Text',
+    url: `/tokens/text`,
+  },
+  {
+    title: 'Width',
+    url: `/tokens/width`,
+  },
+  {
+    title: 'Z-Index',
+    url: `/tokens/z-index`,
+  },
+];
+
+function TokensNav({selected}: Props) {
+  return (
+    <nav className={styles.TokensNav}>
+      <ul>
+        {navItems.map((item) => {
+          if (!item.url) return null;
+          const isCurrent = selected === slugify(item.title);
+          return (
+            <li key={item.title}>
+              <Link
+                href={item.url}
+                aria-current={isCurrent ? 'page' : undefined}
+              >
+                {item.title}
+              </Link>
+            </li>
+          );
+        })}
+      </ul>
+    </nav>
+  );
+}
+
+export default TokensNav;
