+++ polaris.shopify.com/src/components/TokensNav/index.ts
@@ -1,0 +1,3 @@
+import TokensNav from './TokensNav';
+
+export default TokensNav;
