--- polaris.shopify.com/src/components/TokensPage/index.ts
@@ -1,3 +1,0 @@
-import TokensPage from './TokensPage';
-
-export default TokensPage;
